+++ packages/jest-circus/src/legacy-code-todo-rewrite/jestAdapter.ts
@@ -49,9 +49,13 @@ const jestAdapter = async (
     testPath,
   });
 
-  if (config.timers === 'fake') {
+  if (config.timers === 'fake' || config.timers === 'legacy') {
     // during setup, this cannot be null (and it's fine to explode if it is)
     environment.fakeTimers!.useFakeTimers();
+  }
+
+  if (config.timers === 'modern') {
+    environment.fakeTimersModern!.useFakeTimers();
   }
 
   globals.beforeEach(() => {